refactor(deepagents): retarget to the CLI session store and consolidate
Deep Agents has no transcript format — even the Deep Agents CLI persists
sessions as LangGraph SQLite checkpoints — so checkpoint decoding stays,
but the public contract now targets the one standard local store: the
CLI database at ~/.deepagents/sessions.db, where every session is a
thread in the root checkpoint namespace.

API: normalizeCheckpoint now takes just threadId (path defaults to the
CLI store; pythonExecutable still optional). The SDK-era knobs —
required path, checkpointNamespace, checkpointId — are removed; the
latest root-namespace checkpoint is always selected. The Python wrapper
follows: normalize_checkpoint(thread_id, path=None).

Consolidation: all Deep Agents code now lives in src/deepagents.ts
(types, subprocess loader, decoder, and the checkpoint API), replacing
the sprawl across types.ts, adapters/deepagents.ts,
deepagents-checkpoint.ts, and index.ts. finalizeCanonical moved to
canonical.ts where both canonical entry points share it. README
deepagents coverage is collapsed into the single "Deep Agents local
sessions" section.

The fixture DB is regenerated to mirror the CLI layout (multiple
threads, root namespace); integration tests cover per-thread isolation
and Overwrite semantics, and the hermetic SDK test still passes.
